requirements.txt
django==5.1.6
cryptography==44.0.2
reportlab==4.2.5
psycopg[binary]==3.2.6
python-dotenv==1.0.1
gunicorn==23.0.0
Pillow==11.1.0